Dockerized what we have so far
I Added a Dockerfile so that we can run the application
in a container. I also started a docker-compose setup because
we will be hooking this application up to a database real soon
and therefore will have to do this anyway.

I had to set the host value for the flask app to 0.0.0.0 so that
Flask would know that we would try to access from an external source
i.e. from outside of the docker container.

If you do not do this then the application will assume you are calling
it from localhost which you cannot do from inside of a docker container.