Reorganized code/docker to use flask Manager
My modules were acting all weird and it was because
my entire repository of code was in the app module.

I moved out some stuff away from the app folder and
that resolved my import issues.

I still need to fiddle with the docker-compose.yml file
a bit but I should be in good shape to start writing tests
and some actual code.