Added a unit test harness and created new tests
The new tests are for the Plans model. So far
I have the get_by_link working. The other methods
remain to be implemented

Some changes were also made to the existing docker
setup. A new script has been created which will run the
unit tests inside of the docker-compose container set
without having to attach a session to it manually.