Can fetch from API when the page loads up
When the page loads up it will ask the backend
for all data associated with the plan and populate
the Vue object's data fields.

I now need to resolve the discrepancies between the
data model that the backend returns and the data model
that the front-end uses for UI control.