Database updates happen when Vue object changes
Right now we update everything all at once every
time a tiny thing changes. We also run this group
cleaner method over and over again even though there
is a good chance it always returns the same result.

The idea is to eventually send events to the backend
instead of the actual data to store in Mongo.

For example instead of sending the server `POST {items:['a','b']}`
we would say `event - new item 'b' `...or something like that...